requirements.txt
torch==1.13.1
torchvision==0.14.1
numpy
tqdm
tensorboard
matplotlib
pandas

speechbrain~=0.5.15
gdown
datasets
torch_audiomentations
git+https://github.com/Spijkervet/torchaudio-augmentations.git
git+https://github.com/facebookresearch/WavAugment
https://github.com/kpu/kenlm/archive/master.zip
editdistance
wandb
pyctcdecode
torchaudio~=0.13.1
pillow
librosa
hydra-core


